+++ src/main/java/com/gearshiftgaming/se_mod_manager/frontend/view/helper/Dwmapi.java
@@ -1,0 +1,23 @@
+package com.gearshiftgaming.se_mod_manager.frontend.view.helper;
+
+import com.sun.jna.Library;
+import com.sun.jna.Native;
+import com.sun.jna.PointerType;
+import com.sun.jna.platform.win32.WinDef;
+
+
+/**
+ * Copyright (C) 2024 Gear Shift Gaming - All Rights Reserved
+ * You may use, distribute, and modify this code under the terms of the GPL3 license.
+ * <p>
+ * You should have received a copy of the GPL3 license with
+ * this file. If not, please write to: gearshift@gearshiftgaming.com.
+ */
+
+public interface Dwmapi extends Library {
+
+	Dwmapi INSTANCE = Native.load("dwmapi", Dwmapi.class);
+
+	int DwmSetWindowAttribute(WinDef.HWND hwnd, int dwAttribute, PointerType pvAttribute, int cbAttribute);
+
+}
