+++ src/main/resources/view/main-window.fxml
@@ -1,45 +1,10 @@
 <?xml version="1.0" encoding="UTF-8"?>
 
-<?import javafx.geometry.Insets?>
-<?import javafx.scene.control.Button?>
-<?import javafx.scene.control.Label?>
 <?import javafx.scene.layout.AnchorPane?>
 <?import javafx.scene.layout.BorderPane?>
-<?import javafx.scene.layout.HBox?>
-<?import javafx.scene.layout.Region?>
-<?import javafx.scene.layout.StackPane?>
-<?import org.kordamp.ikonli.javafx.FontIcon?>
 
-<AnchorPane minHeight="-Infinity" minWidth="-Infinity" prefHeight="600.0" prefWidth="1050.0" xmlns="http://javafx.com/javafx/21" xmlns:fx="http://javafx.com/fxml/1">
+<AnchorPane fx:id="mainWindowRoot" minHeight="-Infinity" minWidth="-Infinity" prefHeight="600.0" prefWidth="1050.0" xmlns="http://javafx.com/javafx/21" xmlns:fx="http://javafx.com/fxml/1">
     <children>
-      <HBox alignment="CENTER_RIGHT" AnchorPane.bottomAnchor="570.0" AnchorPane.leftAnchor="0.0" AnchorPane.rightAnchor="0.0" AnchorPane.topAnchor="0.0">
-         <children>
-            <Label fx:id="appNameVersion" text="SEMM v-">
-               <padding>
-                  <Insets left="5.0" />
-               </padding>
-            </Label>
-            <Region prefHeight="200.0" prefWidth="200.0" HBox.hgrow="ALWAYS" />
-            <StackPane>
-               <children>
-                  <Button fx:id="minimizeButton" focusTraversable="false" maxHeight="-Infinity" maxWidth="-Infinity" minHeight="-Infinity" minWidth="-Infinity" mnemonicParsing="false" onAction="#minimize" prefHeight="30.0" prefWidth="40.0" />
-                  <FontIcon iconLiteral="codicon-chrome-minimize" mouseTransparent="true" />
-               </children>
-            </StackPane>
-            <StackPane>
-               <children>
-                  <Button fx:id="maximizeRestoreButton" focusTraversable="false" maxHeight="-Infinity" maxWidth="-Infinity" minHeight="-Infinity" minWidth="-Infinity" mnemonicParsing="false" onAction="#maximizeOrRestore" prefHeight="30.0" prefWidth="40.0" />
-                  <FontIcon fx:id="maximizeRestoreIcon" iconLiteral="codicon-chrome-maximize" mouseTransparent="true" />
-               </children>
-            </StackPane>
-            <StackPane>
-               <children>
-                  <Button fx:id="closeButton" focusTraversable="false" maxHeight="-Infinity" maxWidth="-Infinity" minHeight="-Infinity" minWidth="-Infinity" mnemonicParsing="false" onAction="#closeSemm" prefHeight="30.0" prefWidth="40.0" />
-                  <FontIcon iconLiteral="codicon-chrome-close" mouseTransparent="true" />
-               </children>
-            </StackPane>
-         </children>
-      </HBox>
         <BorderPane fx:id="mainWindowLayout" layoutY="27.0" AnchorPane.bottomAnchor="0.0" AnchorPane.leftAnchor="0.0" AnchorPane.rightAnchor="0.0" AnchorPane.topAnchor="30.0">
         </BorderPane>
     </children>
